refactor(eval): lift shared gate constants + stats helpers into stats.go
